Commit 35 - Gauna Nicolas (30/08/2024)

+++ Project_Triathlon/src/model/race/modality/Modality.java
@@ -37,7 +37,7 @@ public abstract class Modality {
 
 
 	//------------------------------------------------>||CLASS METHODS||<--------------------------------------------------------\\
-    public abstract double getRaceTime();
+    public abstract long getRaceTime();
 
 	@Override
 	public String toString() {
